% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_simba_scores.R
\name{add_simba_scores}
\alias{add_simba_scores}
\title{add the SIMBA gene:cell probabilities to an SCE with embeddings available}
\usage{
add_simba_scores(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment instance with \code{pbg_cell_emb} in altExps and \code{g_emb} in metadata}
}
\description{
add the SIMBA gene:cell probabilities to an SCE with embeddings available
}
\examples{
sce = grab_bgdemo_sce()
sce2 = add_simba_scores(sce)
data(p3k.coarse)
nkg7 = as.numeric(assay(altExp(sce2, "sprobs")["NKG7",]))
lm(nkg7~factor(p3k.coarse))
}
