% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{CG_embed_sce}
\alias{CG_embed_sce}
\title{complete processing of SCE}
\usage{
CG_embed_sce(
  sce,
  workdir = tempdir(),
  N_GENES = 1000,
  N_BINS = 5,
  N_PARTITIONS = 1L,
  FEATURIZED = FALSE,
  entity_path = "ents",
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment instance ready for conversion with 'sce_to_triples'}

\item{workdir}{location of temporary storage, defaults to tempdir()}

\item{N_GENES}{numeric(1) passed to 'sce_to_triples'}

\item{N_BINS}{numeric(1) passed to 'sce_to_triples' for discretization}

\item{N_PARTITIONS}{integer(1) for PytorchBigGraph configuration}

\item{FEATURIZED}{logical(1) for PytorchBigGraph configuration}

\item{entity_path}{character(1) subfolder of tempdir for entity recording}

\item{\dots}{passed directly to \code{set_config}}
}
\description{
complete processing of SCE
}
\note{
Any changes to configuration can be passed via ....  For example,
\code{num_epochs} defaults to 5L, pass directly in this function.
}
\examples{
p3k = TENxPBMCData::TENxPBMCData("pbmc3k")
assay(p3k) = as.matrix(assay(p3k)) # dense for now
p3k = scuttle::logNormCounts(p3k)
co = CG_embed_sce(p3k)
co
}
