% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{config_constants}
\alias{config_constants}
\title{produce a list with defaults or substitutes for most configuration parameters}
\usage{
config_constants(
  dimension = 400L,
  init_scale = 0.001,
  max_norm = NULL,
  global_emb = FALSE,
  comparator = "dot",
  bias = FALSE,
  loss_fn = "softmax",
  margin = 0.1,
  regularization_coef = 0.001,
  regularizer = "N3",
  init_path = NULL,
  checkpoint_preservation_interval = NULL,
  num_epochs = 5L,
  num_edge_chunks = NULL,
  max_edges_per_chunk = 1000000000L,
  bucket_order_string = "INSIDE_OUT",
  workers = NULL,
  batch_size = 1000L,
  num_batch_negs = 50L,
  num_uniform_negs = 1000L,
  disable_lhs_negs = FALSE,
  disable_rhs_negs = FALSE,
  lr = 0.1,
  relation_lr = NULL,
  eval_fraction = 0,
  eval_num_batch_negs = 1000L,
  eval_num_uniform_negs = 1000L,
  background_io = FALSE,
  verbose = 0L,
  hogwild_delay = 2,
  dynamic_relations = TRUE,
  num_machines = 1L,
  num_partition_servers = -1L,
  distributed_init_method = NULL,
  distributed_tree_init_order = TRUE,
  num_gpus = 0L,
  num_groups_for_partition_server = 16L,
  num_groups_per_sharded_partition_server = 1L,
  partition_shard_size = 250L,
  half_precision = FALSE
)
}
\arguments{
\item{dimension}{integer(1) defaults to 400}

\item{init_scale}{numeric(1) defaults to 0.001}

\item{max_norm}{defaults to NULL}

\item{global_emb}{logical(1) defaults to FALSE}

\item{comparator}{character(1) defaults to dot}

\item{bias}{logical(1) defaults to FALSE}

\item{loss_fn}{character(1) defaults to softmax}

\item{margin}{numeric(1) defaults to 0.1}

\item{regularization_coef}{numeric(1) defaults to 0.001}

\item{regularizer}{character(1) defaults to N3}

\item{init_path}{defaults to NULL}

\item{checkpoint_preservation_interval}{NULL(1) defaults to NULL}

\item{num_epochs}{integer(1) defaults to 5}

\item{num_edge_chunks}{defaults to NULL}

\item{max_edges_per_chunk}{integer(1) defaults to 1000000000}

\item{bucket_order_string}{character(1) defaults to INSIDE_OUT}

\item{workers}{defaults to NULL}

\item{batch_size}{integer(1) defaults to 1000}

\item{num_batch_negs}{integer(1) defaults to 50}

\item{num_uniform_negs}{integer(1) defaults to 1000}

\item{disable_lhs_negs}{logical(1) defaults to FALSE}

\item{disable_rhs_negs}{logical(1) defaults to FALSE}

\item{lr}{numeric(1) defaults to 0.1}

\item{relation_lr}{defaults to NULL}

\item{eval_fraction}{numeric(1) defaults to 0}

\item{eval_num_batch_negs}{integer(1) defaults to 1000}

\item{eval_num_uniform_negs}{integer(1) defaults to 1000}

\item{background_io}{logical(1) defaults to FALSE}

\item{verbose}{integer(1) defaults to 0}

\item{hogwild_delay}{numeric(1) defaults to 2}

\item{dynamic_relations}{logical(1) defaults to TRUE}

\item{num_machines}{integer(1) defaults to 1}

\item{num_partition_servers}{integer(1) defaults to -1}

\item{distributed_init_method}{defaults to NULL}

\item{distributed_tree_init_order}{logical(1) defaults to TRUE}

\item{num_gpus}{integer(1) defaults to 0}

\item{num_groups_for_partition_server}{integer(1) defaults to 16}

\item{num_groups_per_sharded_partition_server}{integer(1) defaults to 1}

\item{partition_shard_size}{integer(1) defaults to 250}

\item{half_precision}{logical(1) defaults to FALSE}
}
\value{
S3class pbgconf which inherits from "list"
}
\description{
produce a list with defaults or substitutes for most configuration parameters
}
