% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_triples.R
\name{docol}
\alias{docol}
\title{helper function}
\usage{
docol(sce, co, colname = "Barcode", assayname = "disc")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{co}{character()}

\item{colname}{character(1)}

\item{assayname}{character(1)}
}
\description{
helper function
}
