% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_demo.R
\name{grab_bgdemo}
\alias{grab_bgdemo}
\title{cache the demonstration BigGraph output (train only)}
\usage{
grab_bgdemo(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocBigGraphDemo/demo4.zip"
)
}
\arguments{
\item{cache}{BiocFileCache instance or equivalent}

\item{url}{character(1) defaults to OSN address of demo4.zip, produced March 2024}
}
\description{
cache the demonstration BigGraph output (train only)
}
\note{
This function will check for presence of url in cache using bfcquery;
if a hit is found, returns the rpath associated with the last
matching record.  etags can be available for use with bfcneedsupdate.
}
\examples{
ca = BiocFileCache::BiocFileCache()
bgdempath = grab_bgdemo(ca)
bgdempath
}
