% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_demo.R
\name{grab_bgdemo_sce}
\alias{grab_bgdemo_sce}
\title{cache the demonstration BigGraph output (train only)}
\usage{
grab_bgdemo_sce(
  cache = BiocFileCache::BiocFileCache(),
  url =
    "https://mghp.osn.xsede.org/bir190004-bucket01/BiocBigGraphDemo/p3k_e35_g1500_75d.rda",
  use_400 = FALSE
)
}
\arguments{
\item{cache}{BiocFileCache instance or equivalent}

\item{url}{character(1) defaults to OSN address of 1500 gene, 35 epoch
pbmc3k embedding, produced April 2024}

\item{use_400}{logical(1) if TRUE use a slightly outdated
structure including a 400-d embedding, otherwise a 75-d embedding is used}
}
\value{
SingleCellExperiment
}
\description{
cache the demonstration BigGraph output (train only)
}
\note{
This function will check for presence of url in cache using bfcquery.
}
