% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_embeddings.R
\name{ingest_embeddings}
\alias{ingest_embeddings}
\title{work with the output of CG_embed_sce}
\usage{
ingest_embeddings(outemb)
}
\arguments{
\item{outemb}{list derived from configuration schema produced
in the embedding function}
}
\description{
work with the output of CG_embed_sce
}
