% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo4.R
\name{make_demo_emb}
\alias{make_demo_emb}
\title{illustrate working with trained embedding}
\usage{
make_demo_emb(path)
}
\arguments{
\item{path}{location where compressed embedding output is found}
}
\description{
illustrate working with trained embedding
}
\examples{
gg = grab_bgdemo()
td = tempdir()
unzip(gg, exdir=td)
mm = make_demo_emb(td)
mm
rhdf5::h5ls(dir(mm$conf$edge_paths[1], full=TRUE))
}
