% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{make_entity_schema}
\alias{make_entity_schema}
\title{produce EntitySchema, only for use within BasiliskEnvironment}
\usage{
make_entity_schema(
  num_partitions = 1L,
  featurized = FALSE,
  dimension = NULL,
  pbgref
)
}
\arguments{
\item{num_partitions}{integer(1)}

\item{featurized}{logical(1)}

\item{dimension}{NULL Or integer(1)}

\item{pbgref}{instance of torchbiggraph module}
}
\description{
produce EntitySchema, only for use within BasiliskEnvironment
}
\examples{
pbg = reticulate::import("torchbiggraph")
EntitySchema(pbgref = pbg)
}
