% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogini.R
\name{max_scores}
\alias{max_scores}
\title{obtain the simba max metric for genes}
\usage{
max_scores(sce, N = 50)
}
\arguments{
\item{sce}{SingleCellExperiment with simba scores available}

\item{N}{number of cells with highest scores to average}
}
\description{
obtain the simba max metric for genes
}
\examples{
sc3k = grab_bgdemo_sce()
sc3k = add_simba_scores(sc3k)
gi = gini(t(assay(altExps(sc3k)$sprobs)))
ma = as.numeric(max_scores(sc3k))
mydf = data.frame(max=as.numeric(ma), gini=gi, gene=rownames(altExps(sc3k)$sprobs))
litd = mydf[order(mydf$gini,decreasing=TRUE),][1:25,]
hh = mydf[mydf$gene=="GAPDH",] 
litd = rbind(litd, hh)
mm = ggplot(mydf, aes(x=max, y=gini, text=gene)) + geom_point() 
mm + ggrepel::geom_text_repel(data=litd, 
         aes(x=max, y=gini, label=gene), colour="purple")
}
