% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogini.R
\name{norm_scores}
\alias{norm_scores}
\title{obtain the simba norm metric for genes}
\usage{
norm_scores(sce)
}
\arguments{
\item{sce}{SingleCellExperiment with simba scores available}
}
\description{
obtain the simba norm metric for genes
}
