% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p3k.coarse}
\alias{p3k.coarse}
\title{provisional labels from SingleR for 2700 PBMCs}
\format{
character(2700)
}
\usage{
p3k.coarse
}
\description{
provisional labels from SingleR for 2700 PBMCs
}
\note{
Based on HumanPrimaryCellAtlas label.main
}
\keyword{datasets}
