% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p3k.fine}
\alias{p3k.fine}
\title{provisional labels from SingleR for 2700 PBMCs}
\format{
character(2700)
}
\usage{
p3k.fine
}
\description{
provisional labels from SingleR for 2700 PBMCs
}
\keyword{datasets}
