% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo4.R
\name{print.CG_emb}
\alias{print.CG_emb}
\title{printer for configuration}
\usage{
\method{print}{CG_emb}(x, ...)
}
\arguments{
\item{x}{CG_emb instance}

\item{\dots}{not used}
}
\description{
printer for configuration
}
