% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{print.pbgconf}
\alias{print.pbgconf}
\title{printer for long configuration object}
\usage{
\method{print}{pbgconf}(x, ...)
}
\arguments{
\item{x}{inherits from "pbgconf"}

\item{\dots}{not used}
}
\description{
printer for long configuration object
}
