% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{set_config}
\alias{set_config}
\title{set configuration schema parameters}
\usage{
set_config(...)
}
\arguments{
\item{\dots}{named arguments to override elements of \code{config_constants}}
}
\value{
instance of 'pbgconf'
}
\description{
set configuration schema parameters
}
