% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gpu.R
\name{set_gpu}
\alias{set_gpu}
\title{needed to ensure code for gpu usage is compiled}
\usage{
set_gpu()
}
\description{
needed to ensure code for gpu usage is compiled
}
\note{
Runs basilisk-managed pip to compile and install
GPU-enabled torchbiggraph.
}
