% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gpu.R
\name{set_train}
\alias{set_train}
\title{needed to allow subprocess to find torchbiggraph for train module}
\usage{
set_train()
}
\description{
needed to allow subprocess to find torchbiggraph for train module
}
\note{
Runs basilisk-managed pip to compile and install
}
