% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_simba.R
\name{simba_barplot}
\alias{simba_barplot}
\title{produce simba barplot for selected gene}
\usage{
simba_barplot(sce, colvar, sym)
}
\arguments{
\item{sce}{a SingleCellExperiment for which add_simba_scores has been run}

\item{colvar}{character(1) variable in colData(sce) to use for cell type,
defining bar colors}

\item{sym}{character(1) name of row in sce for which expression scores should
be visualized}
}
\value{
ggplot instance
}
\description{
produce simba barplot for selected gene
}
\examples{
sce = grab_bgdemo_sce()
data(p3k.coarse)
sce$celltype = p3k.coarse
sce = add_simba_scores(sce)
simba_barplot(sce, "celltype", "NKG7") +
 ylab("SIMBA celltype:expression score for NKG7") + ggtitle("PBMC3K")
}
