% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{simple_conf}
\alias{simple_conf}
\title{no bas}
\usage{
simple_conf(
  sce,
  workdir = tempdir(),
  N_GENES = 1000,
  N_BINS = 5,
  N_PARTITIONS = 1L,
  FEATURIZED = FALSE,
  entity_path = "ents",
  ...
)
}
\description{
no bas
}
\examples{
p3k = TENxPBMCData::TENxPBMCData("pbmc3k")
co = simple_conf(p3k, dynamic_relations=FALSE)
co
}
