% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barpapp.R
\name{barpapp}
\alias{barpapp}
\title{app for exploring the barplots}
\usage{
barpapp(sce, colvar)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{colvar}{character(1) must be present in colData(sce)}
}
\description{
app for exploring the barplots
}
\examples{
example(simba_barplot) # defines sce, adds 'celltype'
data(p3k.fine)
sce$celltype = p3k.fine
if (interactive()) {
barpapp(sce, "celltype")
}
}
