% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_embeddings.R
\name{ingest_embeddings}
\alias{ingest_embeddings}
\title{work with the output of CG_embed_sce}
\usage{
ingest_embeddings(outemb)
}
\arguments{
\item{outemb}{list derived from configuration schema produced
in the embedding function; for demo4, use \code{make_demo_emb}}
}
\description{
work with the output of CG_embed_sce
}
\examples{
example(make_demo_emb) # creates mm
nnn = ingest_embeddings(mm)
names(nnn)
pp = prcomp(t(nnn$cemb))
data(p3k.fine)
p3k=TENxPBMCData("pbmc3k")
p3k = scuttle::logNormCounts(p3k)
mm = match(p3k$Barcode, nnn$C_entities_ordered)
pairs(pp$x[mm,2:4], pch=19, cex=.4, col=factor(p3k.fine))
}
